# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AGI Unified Framework API
  description: |
    AGI Unified Framework 提供完整的AI系统能力API，包括：
    
    - **核心服务**: LLM推理、记忆管理、工具调用
    - **多模态**: 图像、音频、视频处理
    - **多智能体**: 智能体管理、任务协调
    - **安全**: 审计、沙箱执行、数据保护
    
  version: 1.0.0
  contact:
    name: AGI Framework Team
    email: support@agi-framework.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发服务器
  - url: https://api.agi-framework.org/v1
    description: 生产服务器
...
paths:
...
  /agents/{agent_id}:
    get:
      tags: [agents]
      summary: 获取智能体详情
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: 智能体详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [agents]
      summary: 删除智能体
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    AgentId:
      name: agent_id
      in: path
      required: true
      description: 智能体ID
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, inactive, busy]
        created_at:
          type: string
          format: date-time
        config:
          type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - BearerAuth: []
  - ApiKeyAuth: []